+++ cpp/open3d/geometry/PointCloud.cpp
@@ -71,6 +71,11 @@ OrientedBoundingBox PointCloud::GetOrientedBoundingBox(bool robust) const {
     return OrientedBoundingBox::CreateFromPoints(points_, robust);
 }
 
+OrientedBoundingBox PointCloud::GetMinimalOrientedBoundingBox(
+        bool robust) const {
+    return OrientedBoundingBox::CreateFromPointsMinimal(points_, robust);
+}
+
 PointCloud &PointCloud::Transform(const Eigen::Matrix4d &transformation) {
     TransformPoints(transformation, points_);
     TransformNormals(transformation, normals_);