+++ cpp/open3d/geometry/PointCloud.h
@@ -72,6 +72,8 @@ public:
     AxisAlignedBoundingBox GetAxisAlignedBoundingBox() const override;
     OrientedBoundingBox GetOrientedBoundingBox(
             bool robust = false) const override;
+    OrientedBoundingBox GetMinimalOrientedBoundingBox(
+            bool robust = false) const override;
     PointCloud &Transform(const Eigen::Matrix4d &transformation) override;
     PointCloud &Translate(const Eigen::Vector3d &translation,
                           bool relative = true) override;
@@ -367,6 +369,43 @@ public:
             const int num_iterations = 100,
             const double probability = 0.99999999) const;
 
+    /// \brief Robustly detect planar patches in the point cloud using.
+    /// Araújo and Oliveira, “A robust statistics approach for plane
+    /// detection in unorganized point clouds,” Pattern Recognition, 2020.
+    ///
+    /// \param normal_variance_threshold_deg Planes having point normals with
+    /// high variance are rejected. The default value is 60 deg. Larger values
+    /// would allow more noisy planes to be detected. \param coplanarity_deg The
+    /// curvature of plane detections are scored using the angle between the
+    /// plane's normal vector and an auxiliary vector. An ideal plane would have
+    /// a score of 90 deg. The default value for this threshold is 75 deg, and
+    /// detected planes with scores lower than this are rejected. Large
+    /// threshold values encourage a tighter distribution of points around the
+    /// detected plane, i.e., less curvature. \param outlier_ratio Maximum
+    /// allowable ratio of outliers in associated plane points before plane is
+    /// rejected. \param min_plane_edge_length A patch's largest edge must
+    /// greater than this value to be considered a true planar patch. If set to
+    /// 0, defaults to 1% of largest span of point cloud. \param min_num_points
+    /// Determines how deep the associated octree becomes and how many points
+    /// must be used for estimating a plane. If set to 0, defaults to 0.1% of
+    /// the number of points in point cloud. \param search_param Point neighbors
+    /// are used to grow and merge detected planes. Neighbors are found with
+    /// KDTree search using these params. More neighbors results in higher
+    /// quality patches at the cost of compute. \return Returns a list of
+    /// detected planar patches, represented as OrientedBoundingBox objects,
+    /// with the third column (z) of R indicating the planar patch normal
+    /// vector. The extent in the z direction is non-zero so that the
+    /// OrientedBoundingBox contains the points that contribute to the plane
+    /// detection.
+    std::vector<std::shared_ptr<OrientedBoundingBox>> DetectPlanarPatches(
+            double normal_variance_threshold_deg = 60,
+            double coplanarity_deg = 75,
+            double outlier_ratio = 0.75,
+            double min_plane_edge_length = 0.0,
+            size_t min_num_points = 0,
+            const geometry::KDTreeSearchParam &search_param =
+                    geometry::KDTreeSearchParamKNN()) const;
+
     /// \brief Factory function to create a pointcloud from a depth image and a
     /// camera model.
     ///
@@ -417,13 +456,13 @@ public:
             const Eigen::Matrix4d &extrinsic = Eigen::Matrix4d::Identity(),
             bool project_valid_depth_only = true);
 
-    /// \brief Function to create a PointCloud from a VoxelGrid.
+    /// \brief Factory Function to create a PointCloud from a VoxelGrid.
     ///
     /// It transforms the voxel centers to 3D points using the original point
     /// cloud coordinate (with respect to the center of the voxel grid).
     ///
     /// \param voxel_grid The input VoxelGrid.
-    std::shared_ptr<PointCloud> CreateFromVoxelGrid(
+    static std::shared_ptr<PointCloud> CreateFromVoxelGrid(
             const VoxelGrid &voxel_grid);
 
 public: