+++ cpp/open3d/t/geometry/TriangleMesh.h
@@ -801,6 +801,9 @@ public:
     /// Create an axis-aligned bounding box from vertex attribute "positions".
     AxisAlignedBoundingBox GetAxisAlignedBoundingBox() const;
 
+    /// Create an oriented bounding box from vertex attribute "positions".
+    OrientedBoundingBox GetOrientedBoundingBox() const;
+
     /// Fill holes by triangulating boundary edges.
     ///
     /// This function always uses the CPU device.