+++ cpp/open3d/visualization/utility/PointCloudPicker.h
@@ -52,9 +52,33 @@ public:
     Eigen::Vector3d GetMinBound() const final;
     Eigen::Vector3d GetMaxBound() const final;
     Eigen::Vector3d GetCenter() const final;
+
+    /// If point cloud does not exist, creates an axis-aligned bounding box
+    /// form its default constructor. If point cloud exists, creates the
+    /// axis-aligned bounding box around it. Further details in
+    /// AxisAlignedBoundingBox::CreateFromPoints()
     geometry::AxisAlignedBoundingBox GetAxisAlignedBoundingBox() const final;
+
+    /// If point cloud does not exist, creates an oriented bounding box
+    /// form its default constructor. If point cloud exists, creates the
+    /// oriented bounding box around it. Further details in
+    /// OrientedBoundingBox::CreateFromPoints()
+    /// \param robust If set to true uses a more robust method which works
+    ///               in degenerate cases but introduces noise to the points
+    ///               coordinates.
     geometry::OrientedBoundingBox GetOrientedBoundingBox(
             bool robust = false) const final;
+
+    /// If point cloud does not exist, creates an oriented bounding box
+    /// form its default constructor. If point cloud exists, creates the minimal
+    /// oriented bounding box around it.
+    /// Further details in OrientedBoundingBox::CreateFromPointsMinimal()
+    /// \param robust If set to true uses a more robust method which works
+    ///               in degenerate cases but introduces noise to the points
+    ///               coordinates.
+    geometry::OrientedBoundingBox GetMinimalOrientedBoundingBox(
+            bool robust = false) const final;
+
     PointCloudPicker& Transform(const Eigen::Matrix4d& transformation) override;
     PointCloudPicker& Translate(const Eigen::Vector3d& translation,
                                 bool relative = true) override;