+++ CHANGELOG.md
@@ -6,11 +6,12 @@
 * Corrected documentation for KDTree (typo in Notebook) (PR #4744)
 * Remove `setuptools` and `wheel` from requirements for end users (PR #5020)
 * Fix various typos (PR #5070)
-* Exposed more functionality in SLAM pipeline
+* Exposed more functionality in SLAM and odometry pipelines
 * Fix for depth estimation for VoxelBlockGrid
 * Reserve fragment buffer for VoxelBlockGrid operations
 * Fix raycasting scene: Allow setting of number of threads that are used for building a raycasting scene
 * Fix Python bindings for CUDA device synchronization, voxel grid saving (PR #5425)
+* Support msgpack versions without cmake
 
 ## 0.13